// apps/mcp-app/src/widget/focused/generated-method-map.ts
export type RetKind =
	| 'this'
	| 'shape'
	| 'shape-or-null'
	| 'shapes'
	| 'id'
	| 'id-or-null'
	| 'ids'
	| 'id-set'

export const METHOD_MAP: Record<string, MethodSpec> = {
// ...
	createShape: { args: ['shape-partial'], ret: 'this' },
// ...
}